fix: address all PR review items — wire SummonerDamagedEvent, type-safe enums, DRY helpers, constants
Major fixes:
- Wire SummonerDamagedEvent in SimBehavior.DealSummonerDamage
- Change BattleScene.CurrentState from int to GameState enum
- Fix shop ownership tests via ProfileRepository.LoadProfile(string) overload
- Change BattleSessionConfig.AiPersonality from string to enum
- Replace fabricated GodotObject with null in SummonerVisual MP polling

Minor fixes:
- Remove dead grid toggle from debug_menu.gd
- Extract HP/mana bar color constants, delete dead PLAYER_TEAM_VALUE in game_ui.gd
- Remove empty heal branch, extract DamageUnitsInRadius helper in SimEffects
- Add WinReason constants, replace raw team indices in WinCondition
- Extract tween constants + ApplyHpUpdate helper in SummonerVisual
- Replace raw group strings with GroupIDs constants (C# + GDScript)
- Extract TUTORIAL_ENEMY_SPAWN_POS, use Team enum in battle_dialogue_controller
- Use direct autoload access + localization in multiplayer_lobby
- Extract SpawnPlaneHeight constant, use TeamPlayer in InputCollector
- Add MP client casting signal todo to docs/tracking/todos.md

+++ scripts/csharp/Battle/View/Debug/TestBattleScene.cs
@@ -86,11 +86,11 @@ public partial class TestBattleScene : BattleScene
     /// </summary>
     public new void EndGame(int winnerTeam)
     {
-        if (CurrentState == (int)GameState.GameOver)
+        if (CurrentState == GameState.GameOver)
             return;
 
-        CurrentState = (int)GameState.GameOver;
-        EmitSignal(SignalName.StateChanged, CurrentState);
+        CurrentState = GameState.GameOver;
+        EmitSignal(SignalName.StateChanged, (int)CurrentState);
         EmitSignal(SignalName.GameEnded, winnerTeam);
         GetTree().Paused = true;
